+++ src/pages/privileges/outlets/linixUseCase/styles.ts
