+++ src/pages/privileges/outlets/users/Users.stories.tsx
@@ -1,6 +1,8 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { Users } from "./index";
+import { TokenContext } from "@context/TokenContext";
+import { presente } from "@src/mocks/design/tokensWithReference/presente";
+import { Users } from ".";
 
 const story = {
   components: [Users],
@@ -11,7 +13,15 @@ const story = {
   decorators: [
     (Story: StoryFn) => (
       <BrowserRouter>
-        <Story />
+        <TokenContext.Provider
+          value={{
+            tokenWithRef: presente,
+            handleSubmit: () => {},
+            loading: false,
+          }}
+        >
+          <Story />
+        </TokenContext.Provider>
       </BrowserRouter>
     ),
   ],
