+++ src/routes/privileges.tsx
@@ -6,7 +6,7 @@ import { Users } from "@pages/privileges/outlets/users";
 import { Privileges } from "@pages/privileges";
 import { ErrorPage } from "@components/layout/ErrorPage";
 import { CompleteInvitation } from "@pages/privileges/outlets/users/complete-invitation";
-import { UseCases } from "@src/pages/privileges/outlets/useCases";
+import { LinixUseCase } from "@pages/privileges/outlets/linixUseCase";
 
 function PrivilegesRoutes() {
   return (
@@ -14,7 +14,7 @@ function PrivilegesRoutes() {
       <Route path="/" element={<Privileges />}>
         <Route path="options" element={<PrivilegesOptions />} />
         <Route path="users" element={<Users />} />
-        <Route path="useCases" element={<UseCases />} />
+        <Route path="linixUseCase" element={<LinixUseCase />} />
         <Route path="users/invite" element={<Invite />} />
         <Route path="users/edit/:user_id" element={<EditUser />} />
         <Route
